refactor: move repo dropdown to `RepoDropdown` component, move RepoOption interface and repo `source_type`
