+++ src/lib/stores/repo.ts
@@ -16,3 +16,15 @@ export function setRepoUrl(repo: string) {
     "repoType": getRepoType(repo)
   });
 }
+
+export interface RepoOption {
+  label: string;
+  icon: string;
+  source_type: 0 | 1 | 2;
+}
+
+export const repo_options: RepoOption[] = [
+        { label: "GitHub", icon: "brand-github", source_type: 0 },
+        { label: "GitLab", icon: "brand-gitlab", source_type: 1 },
+        { label: "Local", icon: "folder-code", source_type: 2 },
+    ];
