+++ static/global.css
@@ -154,6 +154,16 @@
     --tint-03: rgba(209, 209, 209, 0.7); /* #D1D1D1 */
 }
 
+.label-primary {
+    color: var(--label-primary) !important;
+}
+.label-secondary {
+    color: var(--label-secondary);
+}
+.label-tertiary {
+    color: var(--label-tertiary);
+}
+
 html, body, #app {
     height: 100%;
     margin: 0;
